+++ src/index.js
@@ -1,7 +1,3 @@
-import State, { setState, getState, bindState, addStateListener, clearStateSubscribers, importJsonToState, exportJsonState, exportStringifiedState, clearState, clearAllState } from './state.js'
-import createElement from './dom.js'
-const Leaf = {
-  State, setState, getState, bindState, createElement
-}
-export default Leaf
-export { State, setState, getState, bindState, createElement, addStateListener, clearStateSubscribers, importJsonToState, exportJsonState, exportStringifiedState, clearState, clearAllState }
+const State = require('./state.js')
+State.createElement = require('./dom.js')
+module.exports = State